refact done

+++ MaslyakBank_Token/src/main/java/system/strategy/AuthTokenStrategy.java
@@ -12,19 +12,19 @@ import system.JwtTokenGenerator;
 
 @Component
 @RequiredArgsConstructor
-public class AuthTokenStrategy implements TokenStrategy{
+public class AuthTokenStrategy implements TokenStrategy<Authentication>{
 
     private final JwtTokenGenerator generator;
     private final UserTokenDAO userTokenDAO;
 
     @Override
-    public String createToken(Object principal) {
-        Authentication auth = (Authentication) principal;
+    public String createToken(Authentication auth) {
         CustomUserDetails userDetails = (CustomUserDetails) auth.getPrincipal();
         UsersTable user = userDetails.getUser();
 
-        userTokenDAO.deleteToken(user.getLogin());
-
+       if (userTokenDAO.findTokenByUser(user)){
+           userTokenDAO.deleteToken(user.getLogin());
+       }
         return generator.generateToken(user, TokenLifetime.AUTHENTICATION, TokenRole.AUTH);
     }
 }
